+++ src/search.ts
@@ -1,3 +1,3 @@
 /** @module ./ */
 export { search } from "./search/search"
-export { type Options, type HitDecorator } from "./search/types"
+export type { Options, HitDecorator } from "./search/types"
